My Resume feature: allow updating text inputs from frontend
For making requests inside redux actions I used 'redux-thunk' and followed this simple tutorial https://www.youtube.com/watch\?v\=QLujoPbqnOs

Note: same as with Redux, I'm sure redux-thunk is not totally necessary, but I wanted to get experience with it. Also, take into account that it's an old approach, as the Redux one is

Note: why can't we use async flow as we usually do? why do we need redux-thunk? It didn't convince me, but this may be the reason https://stackoverflow.com/questions/34570758/why-do-we-need-middleware-for-async-flow-in-redux